+++ stfblender/stf_resources/fallback/json_fallback_component.py
@@ -4,11 +4,9 @@ from typing import Any
 
 from ....stfblender_common import STF_ExportContext, STF_ImportContext, STF_TaskSteps, STF_Category, STF_ComponentResourceBase, STF_Handler_Component, STF_Component_Ref, STFReport, add_component, export_component_base, import_component_base, preserve_component_reference
 from ....stfblender_common.blender_grr import BlenderGRR, construct_blender_grr, resolve_blender_grr
+from ...register_stf_data import blender_types
 from .json_fallback_buffer import STF_FallbackBuffer, decode_buffer, encode_buffer
 from .json_fallback_ui import draw_fallback
-
-
-_blender_property_name = "stf_json_fallback_component"
 
 
 class JsonFallbackComponent(STF_ComponentResourceBase):
@@ -25,7 +23,7 @@ class Handler_JsonFallbackComponent(STF_Handler_Component):
 	stf_type = None # pyright: ignore[reportAssignmentType]
 	stf_category = STF_Category.COMPONENT
 	understood_blender_types = [JsonFallbackComponent]
-	blender_property_name = _blender_property_name
+	blender_property_name = "stf_json_fallback_component"
 	single = False
 
 	@classmethod
@@ -85,73 +83,9 @@ class Handler_JsonFallbackComponent(STF_Handler_Component):
 
 def register():
 	# register wherever components could be possibly added
-	setattr(bpy.types.Action, _blender_property_name, bpy.props.CollectionProperty(type=JsonFallbackComponent))
-	setattr(bpy.types.Armature, _blender_property_name, bpy.props.CollectionProperty(type=JsonFallbackComponent))
-	setattr(bpy.types.Brush, _blender_property_name, bpy.props.CollectionProperty(type=JsonFallbackComponent))
-	setattr(bpy.types.Camera, _blender_property_name, bpy.props.CollectionProperty(type=JsonFallbackComponent))
-	setattr(bpy.types.Collection, _blender_property_name, bpy.props.CollectionProperty(type=JsonFallbackComponent))
-	setattr(bpy.types.Curve, _blender_property_name, bpy.props.CollectionProperty(type=JsonFallbackComponent))
-	setattr(bpy.types.TextCurve, _blender_property_name, bpy.props.CollectionProperty(type=JsonFallbackComponent))
-	setattr(bpy.types.Curves, _blender_property_name, bpy.props.CollectionProperty(type=JsonFallbackComponent))
-	setattr(bpy.types.GreasePencil, _blender_property_name, bpy.props.CollectionProperty(type=JsonFallbackComponent))
-	setattr(bpy.types.Key, _blender_property_name, bpy.props.CollectionProperty(type=JsonFallbackComponent))
-	setattr(bpy.types.Lattice, _blender_property_name, bpy.props.CollectionProperty(type=JsonFallbackComponent))
-	setattr(bpy.types.Library, _blender_property_name, bpy.props.CollectionProperty(type=JsonFallbackComponent))
-	setattr(bpy.types.Light, _blender_property_name, bpy.props.CollectionProperty(type=JsonFallbackComponent))
-	setattr(bpy.types.LightProbe, _blender_property_name, bpy.props.CollectionProperty(type=JsonFallbackComponent))
-	setattr(bpy.types.Mask, _blender_property_name, bpy.props.CollectionProperty(type=JsonFallbackComponent))
-	setattr(bpy.types.Material, _blender_property_name, bpy.props.CollectionProperty(type=JsonFallbackComponent))
-	setattr(bpy.types.Mesh, _blender_property_name, bpy.props.CollectionProperty(type=JsonFallbackComponent))
-	setattr(bpy.types.MetaBall, _blender_property_name, bpy.props.CollectionProperty(type=JsonFallbackComponent))
-	setattr(bpy.types.MovieClip, _blender_property_name, bpy.props.CollectionProperty(type=JsonFallbackComponent))
-	setattr(bpy.types.NodeTree, _blender_property_name, bpy.props.CollectionProperty(type=JsonFallbackComponent))
-	setattr(bpy.types.Object, _blender_property_name, bpy.props.CollectionProperty(type=JsonFallbackComponent))
-	setattr(bpy.types.PaintCurve, _blender_property_name, bpy.props.CollectionProperty(type=JsonFallbackComponent))
-	setattr(bpy.types.Palette, _blender_property_name, bpy.props.CollectionProperty(type=JsonFallbackComponent))
-	setattr(bpy.types.PointCloud, _blender_property_name, bpy.props.CollectionProperty(type=JsonFallbackComponent))
-	setattr(bpy.types.Scene, _blender_property_name, bpy.props.CollectionProperty(type=JsonFallbackComponent))
-	setattr(bpy.types.Screen, _blender_property_name, bpy.props.CollectionProperty(type=JsonFallbackComponent))
-	setattr(bpy.types.Sound, _blender_property_name, bpy.props.CollectionProperty(type=JsonFallbackComponent))
-	setattr(bpy.types.Speaker, _blender_property_name, bpy.props.CollectionProperty(type=JsonFallbackComponent))
-	setattr(bpy.types.Text, _blender_property_name, bpy.props.CollectionProperty(type=JsonFallbackComponent))
-	setattr(bpy.types.Texture, _blender_property_name, bpy.props.CollectionProperty(type=JsonFallbackComponent))
-	setattr(bpy.types.Volume, _blender_property_name, bpy.props.CollectionProperty(type=JsonFallbackComponent))
-	setattr(bpy.types.WindowManager, _blender_property_name, bpy.props.CollectionProperty(type=JsonFallbackComponent))
-	setattr(bpy.types.WorkSpace, _blender_property_name, bpy.props.CollectionProperty(type=JsonFallbackComponent))
-	setattr(bpy.types.World, _blender_property_name, bpy.props.CollectionProperty(type=JsonFallbackComponent))
-
+	for blender_type in blender_types:
+		setattr(blender_type, Handler_JsonFallbackComponent.blender_property_name, bpy.props.CollectionProperty(type=JsonFallbackComponent, name="STF Fallback Components", options=set()))
 
 def unregister():
-	if hasattr(bpy.types.Action, _blender_property_name): delattr(bpy.types.Action, _blender_property_name)
-	if hasattr(bpy.types.Armature, _blender_property_name): delattr(bpy.types.Armature, _blender_property_name)
-	if hasattr(bpy.types.Brush, _blender_property_name): delattr(bpy.types.Brush, _blender_property_name)
-	if hasattr(bpy.types.Camera, _blender_property_name): delattr(bpy.types.Camera, _blender_property_name)
-	if hasattr(bpy.types.Collection, _blender_property_name): delattr(bpy.types.Collection, _blender_property_name)
-	if hasattr(bpy.types.Curve, _blender_property_name): delattr(bpy.types.Curve, _blender_property_name)
-	if hasattr(bpy.types.TextCurve, _blender_property_name): delattr(bpy.types.TextCurve, _blender_property_name)
-	if hasattr(bpy.types.Curves, _blender_property_name): delattr(bpy.types.Curves, _blender_property_name)
-	if hasattr(bpy.types.GreasePencil, _blender_property_name): delattr(bpy.types.GreasePencil, _blender_property_name)
-	if hasattr(bpy.types.Key, _blender_property_name): delattr(bpy.types.Key, _blender_property_name)
-	if hasattr(bpy.types.Lattice, _blender_property_name): delattr(bpy.types.Lattice, _blender_property_name)
-	if hasattr(bpy.types.Library, _blender_property_name): delattr(bpy.types.Library, _blender_property_name)
-	if hasattr(bpy.types.Light, _blender_property_name): delattr(bpy.types.Light, _blender_property_name)
-	if hasattr(bpy.types.LightProbe, _blender_property_name): delattr(bpy.types.LightProbe, _blender_property_name)
-	if hasattr(bpy.types.Mask, _blender_property_name): delattr(bpy.types.Mask, _blender_property_name)
-	if hasattr(bpy.types.Material, _blender_property_name): delattr(bpy.types.Material, _blender_property_name)
-	if hasattr(bpy.types.Mesh, _blender_property_name): delattr(bpy.types.Mesh, _blender_property_name)
-	if hasattr(bpy.types.MetaBall, _blender_property_name): delattr(bpy.types.MetaBall, _blender_property_name)
-	if hasattr(bpy.types.MovieClip, _blender_property_name): delattr(bpy.types.MovieClip, _blender_property_name)
-	if hasattr(bpy.types.NodeTree, _blender_property_name): delattr(bpy.types.NodeTree, _blender_property_name)
-	if hasattr(bpy.types.Object, _blender_property_name): delattr(bpy.types.Object, _blender_property_name)
-	if hasattr(bpy.types.PaintCurve, _blender_property_name): delattr(bpy.types.PaintCurve, _blender_property_name)
-	if hasattr(bpy.types.Palette, _blender_property_name): delattr(bpy.types.Palette, _blender_property_name)
-	if hasattr(bpy.types.PointCloud, _blender_property_name): delattr(bpy.types.PointCloud, _blender_property_name)
-	if hasattr(bpy.types.Scene, _blender_property_name): delattr(bpy.types.Scene, _blender_property_name)
-	if hasattr(bpy.types.Sound, _blender_property_name): delattr(bpy.types.Sound, _blender_property_name)
-	if hasattr(bpy.types.Speaker, _blender_property_name): delattr(bpy.types.Speaker, _blender_property_name)
-	if hasattr(bpy.types.Text, _blender_property_name): delattr(bpy.types.Text, _blender_property_name)
-	if hasattr(bpy.types.Texture, _blender_property_name): delattr(bpy.types.Texture, _blender_property_name)
-	if hasattr(bpy.types.Volume, _blender_property_name): delattr(bpy.types.Volume, _blender_property_name)
-	if hasattr(bpy.types.WindowManager, _blender_property_name): delattr(bpy.types.WindowManager, _blender_property_name)
-	if hasattr(bpy.types.WorkSpace, _blender_property_name): delattr(bpy.types.WorkSpace, _blender_property_name)
-	if hasattr(bpy.types.World, _blender_property_name): delattr(bpy.types.World, _blender_property_name)
+	for blender_type in blender_types:
+		if hasattr(blender_type, Handler_JsonFallbackComponent.blender_property_name): delattr(blender_type, Handler_JsonFallbackComponent.blender_property_name)