+++ stfblender/stf_resources/fallback/json_fallback_data.py
@@ -8,9 +8,6 @@ from .json_fallback_buffer import STF_FallbackBuffer, decode_buffer, encode_buff
 from .json_fallback_ui import draw_fallback
 
 
-_blender_property_name = "stf_json_fallback_data"
-
-
 class JsonFallbackData(STF_DataResourceBase):
 	json: bpy.props.StringProperty(name="Raw Json")
 	referenced_resources: bpy.props.CollectionProperty(type=BlenderGRR, name="Referenced Resources", options=set())
@@ -25,7 +22,7 @@ class Handler_JsonFallbackData(STF_Handler_Data):
 	stf_type = None # pyright: ignore[reportAssignmentType]
 	stf_category = STF_Category.DATA
 	understood_blender_types = [JsonFallbackData]
-	blender_property_name = _blender_property_name
+	blender_property_name = "stf_json_fallback_data"
 
 	@classmethod
 	def draw(cls, layout: bpy.types.UILayout, context: bpy.types.Context, resource_ref: STF_Data_Ref, context_resource: bpy.types.Collection | None, resource: JsonFallbackData):
@@ -86,8 +83,8 @@ class Handler_JsonFallbackData(STF_Handler_Data):
 
 
 def register():
-	setattr(bpy.types.Collection, _blender_property_name, bpy.props.CollectionProperty(type=JsonFallbackData, options=set()))
+	setattr(bpy.types.Collection, Handler_JsonFallbackData.blender_property_name, bpy.props.CollectionProperty(type=JsonFallbackData, options=set()))
 
 def unregister():
-	if hasattr(bpy.types.Collection, _blender_property_name):
-		delattr(bpy.types.Collection, _blender_property_name)
+	if hasattr(bpy.types.Collection, Handler_JsonFallbackData.blender_property_name):
+		delattr(bpy.types.Collection, Handler_JsonFallbackData.blender_property_name)